requirements.txt
streamlit
PyPDF2
python-dotenv
langchain
langchain-community
langchain-pinecone
langchain-core
# langchain-chains
langchain-text-splitters
pinecone-client
sentence-transformers